Provides a "prioritized" aggregate listener
One feature brought up during the WG discussion was how to handle
_prioritization_ or _ordering_ of listeners. Most of the time, this can
be accomplished by implementations, and would be outside the scope of
the specification. This patch provides a potential _implementation_ of a
listener _aggregate_ that provides some minimal ordering functionality.

When talking about ordering, one thing brought up specifically was being
able to say "run this one first" and "run this one last".

As such, this implementation does the following:

- Allows prepending and appending listeners.
- Allows specifying **exactly one** listener to run **first**.
- Allows specifying **exactly one** listener to run **last**.
- When `listen()` is invoked, it creates an internal result aggregate,
  and then calls `listen()` on each listener it aggregates. Listeners
  are invoked in the following order:

  - The listener marked "first", if any.
  - All other listeners, in the order in which they were added,
    respective of `prepend()` and `append()` operations.
  - The listener marked "last", if any.

  When complete, it returns an evolved event with the result aggregate
  as the result.

When multiple attempts to add a listener first or last are made, the
aggregate will raise specific exceptions.