Imports classes from mwop-mk4
Imports all providers (and specialized provider interfaces) from
mwop-mk4, and puts them under the new namespace, as well as the
`StoppableTaskTrait` (formerly the `StoppableEventTrait`).

Removes the `ImmutableDispatcher`, as it's no longer relevant, and
renames `Dispatcher` to `TaskProcessor`, and has it compose a
`ListenerProviderInterface` instance instead.

Imports the `LazyListener`, as well as the utility function for creating
it.