Update to follow spec more closely
As discussed during today's meeting, processors MUST NOT return a
different task instance. They should either throw an exception, or
process an `ErrorEventInterface` and return the original task.

To make this simpler, I modified `ErrorEvent` to extend `Exception` and
implement `Throwable`. As such, it stores the original throwable as the
previous exception (and `getThrowable()` proxies to that method now),
and also exposes the original exception message and code.

Additionally, noticed the following:

- If a task is stopped on receipt, we should not process it and instead
  return it immediately.
- I wasn't pulling listeners from the listener provider -- oops!
- I had an errant `notify()` method in the `ErrorTaskProcessor` that
  needed removal.

I've also updated docblocks to reflect what actually happens.