feat: add and document queueing and container registration facilities
Adds the interface `ListenerShouldQueue`, along with a chapter on queued
listeners.

Also adds the factory `EventDispatcherFactory`, which uses the
`$serviceName` argument as the class to instantiate, and passes the
`Psr\EventDispatcher\ListenerProviderInterface` service as a constructor
argument. The various listener provider implementations (except the
`ListenerProviderAggregate`) are wired as invokables in the
`ConfigProvider`, while the `EventDispatcher` and
`ErrorEmittingDispatcher` are each wired to the
`EventDispatcherFactory`.